// frontend/src/app/major/+state/enums/major-status.enum.ts
export enum MajorStatusEnum {
  pending = 'pending',
  loading = 'loading',
  loadSuccess = 'loadSuccess',
  loadError = 'loadError',
  loadDetailsSuccess = 'loadDetailsSuccess',
  loadDetailsFailure = 'loadDetailsFailure',
  createSuccess = 'createSuccess',
  createFailure = 'createFailure',
  updateSuccess = 'updateSuccess',
  updateFailure = 'updateFailure',
  deleteSuccess = 'deleteSuccess',
  deleteFailure = 'deleteFailure',
  approveSuccess = 'approveSuccess',
  approveFailure = 'approveFailure',
}